fix(rename): DoubleDounce -> DoubleBounce
